@@ -1,40 +1,31 @@
+require 'arxiv_import'
+
 class PapersController < ApplicationController
   include PapersHelper
 
   def show
-    @paper = Paper.find_by_uid(params[:id])
+    @paper = Paper.where(uid: Arxiv.strip_version(params[:paper_uid]))
+      .select('id', 'uid', 'title', 'abstract', 'scites_count', 'abs_url', 'pdf_url', 'update_date', 'submit_date', 'pubdate', 'author_comments', 'msc_class', 'report_no', 'journal_ref', 'doi', 'proxy', 'updated_at', 'author_str', 'versions_count')
+      .first!
 
-    if @paper.nil?
-      # Might be a versioned link
-      @paper = Paper.find_by_uid!(params[:id].split(/v\d/)[0])
-      redirect_to paper_path(@paper)
+    @feeds = @paper.feeds
+      .select('fullname', 'uid')
+
+    @authors = @paper.authors
+      .select('fullname', 'searchterm')
+
+    @sciters = @paper.sciters
+      .select('fullname', 'username')
+
+    @scited_by_uid = if current_user && @sciters.find { |s| s.username == current_user.username }
+      { @paper.uid => true }
+    else
+      {}
     end
 
-    @scited = current_user && current_user.scited_papers.where(id: @paper.id).exists?
+    @comments = find_comments_sorted_by_rating
 
-    # Less naive statistical comment sorting as per
-    # http://www.evanmiller.org/how-not-to-sort-by-average-rating.html
-    @toplevel_comments = Comment.find_by_sql([
-      "SELECT *, COALESCE(((cached_votes_up + 1.9208) / NULLIF(cached_votes_up + cached_votes_down, 0) - 1.96 * SQRT((cached_votes_up * cached_votes_down) / NULLIF(cached_votes_up + cached_votes_down, 0) + 0.9604) / NULLIF(cached_votes_up + cached_votes_down, 0)) / (1 + 3.8416 / NULLIF(cached_votes_up + cached_votes_down, 0)), 0) AS ci_lower_bound FROM comments WHERE paper_uid = ? AND ancestor_id IS NULL AND (hidden = FALSE OR user_id = ?) ORDER BY ci_lower_bound DESC, created_at ASC;",
-      @paper.uid,
-      current_user ? current_user.id : nil
-    ])
-
-    @comment_tree = {}
-    @paper.comments.where("ancestor_id IS NOT NULL").order("created_at ASC").each do |c|
-      @comment_tree[c.ancestor_id] ||= []
-      @comment_tree[c.ancestor_id] << c
-    end
-
-    @comments = []
-    @toplevel_comments.each do |c|
-      @comments << c
-      @comments += @comment_tree[c.id]||[]
-    end
-
-    @comments = @comments.reject do |c| 
-      c.deleted && (@comment_tree[c.id].nil? || @comment_tree[c.id].all? { |c2| c2.deleted })
-    end
+    render 'papers/show'
   end
 
   def __quote(val)
@@ -46,25 +37,31 @@ class PapersController < ApplicationController
     advanced = params[:advanced]
     page = params[:page] ? params[:page].to_i : 1
 
-    @search = Paper::Search.new(basic, advanced)
+    @search = Search::Paper::Query.new(basic, advanced)
 
     per_page = 70
 
     if !@search.query.empty?
-      paper_uids = @search.run(from: (page-1)*per_page, size: per_page).documents.map(&:_id)
+      res = @search.run(from: (page-1)*per_page, size: per_page)
 
-      @papers = Paper.includes(:authors, :feeds)
-                     .where(uid: paper_uids)
-                     .index_by(&:uid)
-                     .slice(*paper_uids)
-                     .values
+      papers_by_uid = map_models :uid, Paper.where(uid: res.documents.map(&:_id))
+
+      @papers = res.documents.map do |doc|
+        paper = papers_by_uid[doc[:_id]]
+
+        paper.authors_fullname = doc.authors_fullname
+        paper.authors_searchterm = doc.authors_searchterm
+        paper.feed_uids = doc.feed_uids
+
+        paper
+      end
 
       @pagination = WillPaginate::Collection.new(page, per_page, @search.results.raw.hits.total)
 
       # Determine which folder we should have selected
       @folder_uid = @search.feed && (@search.feed.parent_uid || @search.feed.uid)
 
-      @scited_ids = current_user.scited_papers.pluck(:id) if current_user
+      @scited_by_uid = current_user.scited_by_uid(@papers) if current_user
     end
 
     render :search
@@ -72,7 +69,7 @@ class PapersController < ApplicationController
 
   # Show the users who scited this paper
   def scites
-    @paper = Paper.find_by_uid!(params[:id])
+    @paper = Paper.find_by_uid!(params[:paper_uid])
   end
 
   def next
@@ -122,5 +119,53 @@ class PapersController < ApplicationController
     end
 
     redirect_to papers_path(params.merge(date: pdate, action: nil))
+  end
+
+  private
+
+  def paper_id
+    if has_versioning_suffix?(params[:id])
+      params[:id].split(/v\d/)[0]
+    else
+      params[:id]
+    end
+  end
+
+  def has_versioning_suffix?(id)
+    id =~ /v\d/
+  end
+
+  # Less naive statistical comment sorting as per
+  # http://www.evanmiller.org/how-not-to-sort-by-average-rating.html
+  def find_top_level_comments
+    total_votes = %Q{ NULLIF(cached_votes_up + cached_votes_down, 0) }
+    Comment.find_by_sql([
+      %Q{
+        SELECT *, COALESCE(
+          ((cached_votes_up + 1.9208) / #{total_votes} - 1.96 * SQRT((cached_votes_up * cached_votes_down) / #{total_votes} + 0.9604) / #{total_votes} ) / (1 + 3.8416 / #{total_votes})
+        , 0) AS ci_lower_bound
+        FROM comments
+        WHERE paper_uid = ? AND ancestor_id IS NULL AND deleted = FALSE
+        AND (hidden = FALSE OR user_id = ?)
+        ORDER BY ci_lower_bound DESC, created_at ASC;
+      }, @paper.uid, current_user.try(:id)])
+  end
+
+  def find_comments_with_ancestors(ancestors)
+    ancestor_ids = ancestors.map(&:id)
+    @paper.comments.where(ancestor_id: ancestor_ids, deleted: false).order('created_at ASC')
+  end
+
+  def find_comments_sorted_by_rating
+    toplevel_comments = find_top_level_comments
+    comment_tree = find_comments_with_ancestors(toplevel_comments).group_by(&:ancestor_id)
+
+    comments = []
+    toplevel_comments.each do |ancestor|
+      comments << ancestor
+      comments += comment_tree[ancestor.id] || []
+    end
+
+    comments
   end
 end
