@@ -14,5 +14,5 @@ class CommentReport < ActiveRecord::Base
   belongs_to :comment
 
   # User can't report a comment multiple times
-  validates_uniqueness_of :comment_id, :scope => :user_id
+  validates :comment_id, uniqueness: { scope: :user_id }
 end
