@@ -19,8 +19,15 @@
 #  subscriptions_count    :integer          default(0)
 #  expand_abstracts       :boolean          default(FALSE)
 #  account_status         :text             default("user")
-#  username               :text
+#  username               :text             not null
 #
+
+require 'open-uri'
+require 'ostruct'
+require 'data_helpers'
+
+class Activity < OpenStruct
+end
 
 class User < ActiveRecord::Base
   STATUS_ADMIN = 'admin'
@@ -29,14 +36,20 @@ class User < ActiveRecord::Base
   STATUS_SPAM = 'spam'
   ACCOUNT_STATES = [STATUS_ADMIN, STATUS_MODERATOR, STATUS_USER, STATUS_SPAM]
 
-  has_secure_password
+  # We don't use the default validations because if users sign up
+  # with oauth they don't need a password.
+  has_secure_password(validations: false)
+  attr_accessor :password_confirmation
 
   has_many :scites, dependent: :destroy
   has_many :scited_papers, through: :scites, source: :paper
+  has_many :authorships, dependent: :destroy, class_name: 'Authorship'
+  has_many :authored_papers, through: :authorships, source: :paper
   has_many :comments, -> { order('created_at DESC') }, dependent: :destroy
   has_many :subscriptions, dependent: :destroy
+  has_many :feed_preferences, dependent: :destroy
+  has_many :auth_links, dependent: :destroy
   has_many :feeds, through: :subscriptions
-  has_many :feed_preferences
 
   validates :fullname, presence: true, length: { maximum: 50 }
 
@@ -45,17 +58,37 @@ class User < ActiveRecord::Base
                     uniqueness: { case_sensitive: false }
 
   valid_username_regex = /\A[a-zA-Z0-9\-_\.]+\z/i
-  validates :username, presence: true, 
+  validates :username, presence: true,
             format: { with: valid_username_regex, message: "may only contain alphanumeric characters and - or _" },
                     uniqueness: { case_sensitive: false }
 
+  valid_aid_regex = /\A[a-z0-9_]+\z/
+  validates :author_identifier,
+            format: { with: valid_aid_regex, message: "must be a valid arXiv author id" },
+            allow_blank: true
+
   validate do |user|
+    # Only need a password if it's not oauth
+    needs_password = user.auth_links.empty? && user.password_digest.nil?
+
+    if needs_password && user.password.nil?
+      user.errors.add :password, "must be present"
+    end
+
     if user.password && user.password.length < 6
       user.errors.add :password, "must be at least 6 characters"
     end
 
+    if user.password && user.password_confirmation != user.password
+      user.errors.add :password, "must match password confirmation"
+    end
+
     if user.username && Settings::RESERVED_USERNAMES.include?(user.username.downcase)
       user.errors.add :username, "is already taken"
+    end
+
+    if !user.url.empty? && !user.url.match(/http(s)?:\/\//)
+      user.url = "http://" + user.url
     end
   end
 
@@ -65,6 +98,13 @@ class User < ActiveRecord::Base
     # Reset the user session if vital information changes
     if new_record? || password_digest_changed? || email_changed?
       generate_token(:remember_token)
+    end
+
+    if author_identifier_changed?
+      authorships.destroy_all
+      self.papers_count = 0
+
+      update_authorship!(true) unless author_identifier.empty?
     end
 
     # Switching to/from spam needs propagation to comments
@@ -88,7 +128,11 @@ class User < ActiveRecord::Base
   end
 
   def self.default_username(fullname)
-    "#{fullname.parameterize}"
+    if User.where(username: fullname.parameterize).exists?
+      "#{fullname.parameterize}" + "-#{User.count}"
+    else
+      fullname.parameterize
+    end
   end
 
   def scited?(paper)
@@ -98,27 +142,35 @@ class User < ActiveRecord::Base
   def scite!(paper)
     unless scites.find_by_paper_uid(paper.uid)
       scites.create!(paper_uid: paper.uid)
-      paper.scites_count += 1
-      paper.save
     end
   end
 
   def unscite!(paper)
-    scites.find_by_paper_uid(paper.uid).destroy
-    paper.scites_count -= 1
-    paper.save
+    scites.where(paper: paper.uid).destroy_all
   end
 
-  def subscribed?(feed)
-    subscriptions.find_by_feed_uid(feed.uid)
+  def refresh_scites_count!
+    self.scites_count = Scite.where(user_id: id).count
+    save!
+  end
+
+  # Given a collection of papers, map uids to scite status
+  def scited_by_uid(papers)
+    map_exists :paper_uid, scites.where(paper_uid: papers.map(&:uid))
   end
 
   def subscribe!(feed)
-    subscriptions.create!(feed_uid: feed.uid)
+    unless subscriptions.find_by_feed_uid(feed.uid)
+      subscriptions.create!(feed_uid: feed.uid)
+    end
   end
 
   def unsubscribe!(feed)
     subscriptions.find_by_feed_uid(feed.uid).destroy
+  end
+
+  def subscribed?(feed)
+    subscriptions.find_by_feed_uid(feed.uid)
   end
 
   def has_subscriptions?
@@ -151,7 +203,7 @@ class User < ActiveRecord::Base
     UserMailer.email_change(self, address).deliver
   end
 
-  def active?
+  def email_confirmed?
     active
   end
 
@@ -188,6 +240,81 @@ class User < ActiveRecord::Base
       email: self.email,
       expand_abstracts: self.expand_abstracts
     }.to_json
+  end
+
+  # Given an author_identifier, queries arXiv for papers
+  # authored by this user and makes corresponding Authorship
+  # connections
+  def update_authorship!(saving=false)
+    url = "http://export.arxiv.org/fb/feed/#{author_identifier}/?format=xml"
+    doc = Nokogiri(open(url))
+    doc.css('entry id').each do |el|
+      uid = el.text.match(/arxiv.org\/abs\/(.+)/)[1]
+      uid = Arxiv.strip_version(uid)
+      unless authorships.where(paper_uid: uid).exists?
+        puts "New paper published by #{username}: arxiv/#{uid}"
+        authorships.create(paper_uid: uid)
+      end
+    end
+
+    self.papers_count = Authorship.where(user_id: id).count
+    self.save! unless saving
+  end
+
+  def activity_feed(n=25)
+    # Fancy activity query
+    # We don't need no denormalization here
+    results = execute(%Q{
+      SELECT 'scite' AS event, 'paper' AS subject_type, paper_uid AS subject_id, created_at
+        FROM scites
+        WHERE scites.user_id = ?
+        UNION ALL
+      SELECT 'subscribe' AS event, 'feed' AS subject_type, feed_uid AS subject_id, created_at
+        FROM subscriptions
+        WHERE subscriptions.user_id = ?
+        UNION ALL
+      SELECT 'comment' AS event, 'comment' AS subject_type, id::text AS subject_id, created_at
+        FROM comments
+        WHERE comments.user_id = ? AND comments.deleted IS FALSE
+        UNION ALL
+      SELECT 'authorship' AS event, 'paper' AS subject_type, paper_uid AS subject_id, created_at
+        FROM authorships
+        WHERE authorships.user_id = ?
+      ORDER BY created_at DESC
+      LIMIT ?
+    }, id, id, id, id, n)
+
+    subject_ids = {}
+
+    results.each do |result|
+      type = result['subject_type']
+      if type == 'comment'
+        result['subject_id'] = result['subject_id'].to_i
+      end
+      subject_ids[type] ||= []
+      subject_ids[type] << result['subject_id']
+    end
+
+    subjects = {
+      'paper' => (map_models :uid, Paper.where(uid: subject_ids['paper'])),
+      'feed' => (map_models :uid, Feed.where(uid: subject_ids['feed'])),
+      'comment' => (map_models :id, Comment.where(id: subject_ids['comment']).includes(:paper))
+    }
+
+    activities = results.map do |result|
+      subject = subjects[result['subject_type']][result['subject_id']]
+
+      Activity.new(event: result['event'],
+                   created_at: result['created_at'],
+                   result['subject_type'] => subject)
+    end
+
+    if activities.length < n
+      activities << Activity.new(event: 'signup',
+                                 created_at: created_at)
+    end
+
+    activities
   end
 
   private
