@@ -14,9 +14,13 @@
 //= require jquery_ujs
 //= require underscore
 //= require bootstrap
-//= require jquery.pagedown-bootstrap.combined.min.js
+//= require jquery.pagedown-bootstrap.combined.js
 //= require moment.min.js
 //= require typeahead.bundle.js
 //= require backbone
 //= require datepicker.js
+//= require ZeroClipboard.js
+//= require turbolinks
+//= require nprogress.js
+//= require jquery.timeago.js
 //= require_tree .
