@@ -2,7 +2,7 @@ class PasswordResetsController < ApplicationController
   def create
     user = User.find_by_email(params[:email])
 
-    if (user && user.active?)
+    if user
       user.send_password_reset
       flash[:success] = "Email sent with password reset instructions."
       redirect_to root_url
