@@ -29,12 +29,16 @@
 class Paper < ActiveRecord::Base
   has_many  :versions, -> { order("position ASC") }, dependent: :delete_all,
             foreign_key: :paper_uid, primary_key: :uid
-  has_many  :categories, -> { order("position ASC") }, dependent: :delete_all,
+  has_many  :categories, -> { order("categories.position ASC") }, dependent: :delete_all,
             foreign_key: :paper_uid, primary_key: :uid
   has_many  :authors, -> { order("position ASC") }, dependent: :delete_all,
             foreign_key: :paper_uid, primary_key: :uid
 
-  has_many  :feeds, -> { order("categories.position ASC") }, through: :categories
+  has_many :authorships, dependent: :destroy,
+            foreign_key: :paper_uid, primary_key: :uid
+  has_many :claimants, through: :authorships, source: :user
+
+  has_many  :feeds, through: :categories
 
   has_many  :scites, dependent: :delete_all,
             foreign_key: :paper_uid, primary_key: :uid
@@ -51,8 +55,11 @@ class Paper < ActiveRecord::Base
 
   validate :update_date_is_after_submit_date
 
+  # Re-index after a scite or comment
   after_save do
-    ::Search::Paper.index(self)
+    if scites_count_changed? || comments_count_changed?
+      ::Search::Paper.index(self)
+    end
   end
 
   # Given when a paper was submitted, estimate the
@@ -85,12 +92,17 @@ class Paper < ActiveRecord::Base
       hidden: false
     ).count
 
-    save
+    save!
   end
 
   def refresh_scites_count!
     self.scites_count = Scite.where(paper_uid: uid).count
-    save
+    save!
+  end
+
+  def refresh_versions_count!
+    self.versions_count = Version.where(paper_uid: uid).count
+    save!
   end
 
   def to_param
@@ -101,6 +113,40 @@ class Paper < ActiveRecord::Base
     update_date > submit_date
   end
 
+  def to_bibtex
+    props = {
+      author: author_str.gsub(/\. /, ".~"), # unbreakable space
+      title: title.gsub(/([A-Z]+)/, "{\\1}"),
+      year: pubdate.year,
+      eprint: uid
+    }
+
+    props[:howpublished] = journal_ref unless journal_ref.nil?
+    props[:doi] = doi unless doi.nil?
+    props[:note] = "arXiv:#{uid}v#{versions_count}"
+
+    props = props.map { |k,v| "#{k} = {#{v}}" }
+
+    %Q{@misc{#{uid},
+  #{props.join(",\n  ")}
+}}
+  end
+
+  # For compatibility with search document papers
+  attr_accessor :authors_fullname, :authors_searchterm, :feed_uids
+
+  def authors_fullname
+    @authors_fullname ||= authors.map(&:fullname)
+  end
+
+  def authors_searchterm
+    @authors_searchterm ||= authors.map(&:searchterm)
+  end
+
+  def feed_uids
+    @feed_uids ||= categories.map(&:feed_uid)
+  end
+
   private
     def update_date_is_after_submit_date
       return unless submit_date and update_date
@@ -109,185 +155,5 @@ class Paper < ActiveRecord::Base
         errors.add(:update_date, "must not be earlier than submit_date")
       end
     end
-end
-
-class Paper::Search
-  attr_reader :results
-  attr_accessor :query, :basic, :advanced
-  attr_accessor :conditions, :feed, :authors, :order, :order_sql
-
-  # Split query on non-paren enclosed spaces
-  def psplit(query)
-    split = []
-    depth = 0
-    current = ""
-
-    query.chars.each_with_index do |ch, i|
-      if i == query.length-1
-        split << current+ch
-      elsif ch == ' ' && depth == 0
-        split << current
-        current = ""
-      else
-        current << ch
-
-        if ch == '('
-          depth += 1
-        elsif ch == ')'
-          depth -= 1
-        end
-      end
-    end
-
-    split
-  end
-
-  # Strip field prefix
-  def tstrip(term)
-    ['au:','ti:','abs:','in:','order:','date:'].each do |prefix|
-      term = term.split(':', 2)[1] if term.start_with?(prefix)
-    end
-
-    #if term[0] == '(' && term[-1] == ')'
-    #  term[1..-2]
-    #else
-    term
-    #end
-  end
-
-  def parse_date(term)
-    if term.match(/^\d\d\d\d$/)
-      Chronic.parse(term+'-01-01')
-    elsif term.match(/^\d\d\d\d-\d\d$/)
-      Chronic.parse(term+'-01')
-    else
-      Chronic.parse(term)
-    end
-  end
-
-  def parse_date_range(term)
-    if term.include?('..')
-      first, last = term.split('..').map { |t| parse_date(t) }
-      first ||= 1000.years.ago
-      last ||= Time.now
-      first..last
-    else
-      # Allow implicit ranges like date:2012
-      time = parse_date(term)
-      if term.match(/^\d\d\d\d$/)
-        time.beginning_of_year..time.end_of_year
-      elsif term.match(/^\d\d\d\d-\d\d$/)
-        time.beginning_of_month..time.end_of_month
-      else
-        time.beginning_of_day..time.end_of_day
-      end
-    end
-  end
-
-  def initialize(basic, advanced)
-    @basic = basic
-    @advanced = advanced
-
-    @query = [@basic, @advanced].join(' ').strip
-
-    @general = nil # Term to apply as OR across all text fields
-    @conditions = {}
-    @authors = []
-    @date_range = nil
-    @orders = []
-
-    psplit(@query).each do |term|
-      if term.start_with?('au:')
-        if term.include?('_')
-          @authors << tstrip(term)
-          @conditions[:authors_searchterm] ||= []
-          @conditions[:authors_searchterm] << tstrip(term)
-        else
-          @authors << tstrip(term)
-          @conditions[:authors_fullname] ||= []
-          @conditions[:authors_fullname] << tstrip(term)
-        end
-      elsif term.start_with?('ti:')
-        @conditions[:title] ||= []
-        @conditions[:title] << tstrip(term)
-      elsif term.start_with?('abs:')
-        @conditions[:abstract] ||= []
-        @conditions[:abstract] << tstrip(term)
-      elsif term.start_with?('in:')
-        @conditions[:feed_uids] ||= []
-        @conditions[:feed_uids] << tstrip(term)
-      elsif term.start_with?('order:')
-        @orders << tstrip(term).to_sym
-      elsif term.start_with?('date:')
-        @date_range = parse_date_range(tstrip(term))
-      else
-        if @general
-          @general += ' ' + term
-        else
-          @general = term
-        end
-      end
-    end
-
-    @sort = []
-
-    @orders = [:scites] if @orders.empty?
-
-    @orders.each do |order|
-      case order
-      when :scites then @sort << { scites_count: 'desc' }
-      when :comments then @sort << { comments_count: 'desc' }
-      when :recency then @sort << { pubdate: 'desc' }
-      when :relevancy then nil # Standard text match sort
-      end
-    end
-
-    # Everything is post-sorted by pubdate except :relevancy
-    unless @sort.empty? || @orders.include?(:recency)
-      @sort << { pubdate: 'desc' }
-    end
-  end
-
-  def run(opts={})
-    es_query = []
-    es_query << @general unless @general.nil?
-    @conditions.each do |cond, vals|
-      vals.each do |val|
-        es_query << "#{cond}:#{val}"
-      end
-    end
-
-    p es_query.join(' ')
-
-    filter = if @date_range
-      {
-        range: {
-          pubdate: {
-            from: @date_range.first,
-            to: @date_range.last
-          }
-        }
-      }
-    else
-      nil
-    end
-
-    params = {
-      sort: @sort,
-      query: {
-        filtered: {
-          query: {
-            query_string: {
-              query: es_query.join(' '),
-              default_operator: 'AND'
-            }
-          },
-          filter: filter
-        }
-      }
-    }.merge(opts)
-
-    @results = ::Search::Paper.find(params)
-  end
 end
 
