@@ -1,4 +1,4 @@
-SciRate3::Application.configure do
+SciRate::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
 
   # In the development environment your application's code is reloaded on
@@ -6,9 +6,12 @@ SciRate3::Application.configure do
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
 
-  # Show full error reports and disable caching
+  # Do not eager load code on boot.
+  config.eager_load = false
+
+  # Show full error reports
   config.consider_all_requests_local       = true
-  config.action_controller.perform_caching = true
+  config.action_controller.perform_caching = true # false
 
   # Don't care if the mailer can't send
   config.action_mailer.raise_delivery_errors = false
@@ -16,17 +19,20 @@ SciRate3::Application.configure do
   # Print deprecation notices to the Rails logger
   config.active_support.deprecation = :log
 
-  # Only use best-standards-support built into browsers
-  config.action_dispatch.best_standards_support = :builtin
+  # Raise an error on page load if there are pending migrations.
+  config.active_record.migration_error = :page_load
 
-  # Do not compress assets
-  config.assets.compress = false
-
-  # Expands the lines which load the assets
+  # Debug mode disables concatenation and preprocessing of assets.
+  # This option may cause significant delays in view rendering with a large
+  # number of complex assets.
   config.assets.debug = true
 
-  config.eager_load = false
+  # Adds additional error checking when serving assets at runtime.
+  # Checks for improperly declared sprockets dependencies.
+  # Raises helpful error messages.
+  config.assets.raise_runtime_errors = true
 
+  # Debug mail alternative to sendgrid
   ActionMailer::Base.delivery_method = :smtp
   ActionMailer::Base.perform_deliveries = true
   ActionMailer::Base.raise_delivery_errors = true
@@ -41,4 +47,6 @@ SciRate3::Application.configure do
   }
 
   WillPaginate.per_page = 100
+
+  config.cache_store = :null_store # :memory_store
 end
