@@ -1,2 +1,2 @@
-SciRate3::Application.config.secret_key_base = Settings::SECRET_KEY_BASE
-SciRate3::Application.config.secret_token = Settings::SECRET_TOKEN
+SciRate::Application.config.secret_key_base = Settings::SECRET_KEY_BASE
+SciRate::Application.config.secret_token = Settings::SECRET_TOKEN
