@@ -11,10 +11,21 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20140305044305) do
+ActiveRecord::Schema.define(version: 20140622150016) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
+
+  create_table "auth_links", force: true do |t|
+    t.string   "provider",         null: false
+    t.string   "uid",              null: false
+    t.string   "oauth_token",      null: false
+    t.datetime "oauth_expires_at", null: false
+    t.integer  "user_id",          null: false
+    t.datetime "created_at"
+    t.datetime "updated_at"
+    t.text     "auth",             null: false
+  end
 
   create_table "authors", force: true do |t|
     t.integer "position",   null: false
@@ -26,6 +37,13 @@ ActiveRecord::Schema.define(version: 20140305044305) do
   add_index "authors", ["paper_uid"], name: "index_authors_on_paper_uid", using: :btree
   add_index "authors", ["position", "paper_uid"], name: "index_authors_on_position_and_paper_uid", unique: true, using: :btree
   add_index "authors", ["searchterm"], name: "index_authors_on_searchterm", using: :btree
+
+  create_table "authorships", force: true do |t|
+    t.integer  "user_id",    null: false
+    t.text     "paper_uid",  null: false
+    t.datetime "created_at"
+    t.datetime "updated_at"
+  end
 
   create_table "categories", force: true do |t|
     t.integer  "position",                                       null: false
@@ -52,33 +70,54 @@ ActiveRecord::Schema.define(version: 20140305044305) do
   add_index "comment_reports", ["user_id", "comment_id"], name: "index_comment_reports_on_user_id_and_comment_id", unique: true, using: :btree
 
   create_table "comments", force: true do |t|
-    t.integer  "user_id",                           null: false
-    t.integer  "score",             default: 0,     null: false
-    t.integer  "cached_votes_up",   default: 0,     null: false
-    t.integer  "cached_votes_down", default: 0,     null: false
-    t.boolean  "hidden",            default: false, null: false
+    t.integer  "user_id",                            null: false
+    t.integer  "score",              default: 0,     null: false
+    t.integer  "cached_votes_up",    default: 0,     null: false
+    t.integer  "cached_votes_down",  default: 0,     null: false
+    t.boolean  "hidden",             default: false, null: false
     t.integer  "parent_id"
     t.integer  "ancestor_id"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.text     "content",                           null: false
-    t.boolean  "deleted",           default: false, null: false
-    t.text     "paper_uid",         default: "",    null: false
+    t.text     "content",                            null: false
+    t.boolean  "deleted",            default: false, null: false
+    t.text     "paper_uid",          default: "",    null: false
+    t.boolean  "hidden_from_recent", default: false, null: false
   end
 
   add_index "comments", ["ancestor_id"], name: "index_comments_on_ancestor_id", using: :btree
+  add_index "comments", ["deleted"], name: "index_comments_on_deleted", using: :btree
+  add_index "comments", ["hidden"], name: "index_comments_on_hidden", using: :btree
+  add_index "comments", ["hidden_from_recent"], name: "index_comments_on_hidden_from_recent", using: :btree
+  add_index "comments", ["id", "paper_uid", "deleted", "hidden", "hidden_from_recent"], name: "index_comments_for_recent", using: :btree
   add_index "comments", ["paper_uid"], name: "index_comments_on_paper_uid", using: :btree
   add_index "comments", ["parent_id"], name: "index_comments_on_parent_id", using: :btree
   add_index "comments", ["user_id"], name: "index_comments_on_user_id", using: :btree
 
+  create_table "delayed_jobs", force: true do |t|
+    t.integer  "priority",   default: 0, null: false
+    t.integer  "attempts",   default: 0, null: false
+    t.text     "handler",                null: false
+    t.text     "last_error"
+    t.datetime "run_at"
+    t.datetime "locked_at"
+    t.datetime "failed_at"
+    t.string   "locked_by"
+    t.string   "queue"
+    t.datetime "created_at"
+    t.datetime "updated_at"
+  end
+
+  add_index "delayed_jobs", ["priority", "run_at"], name: "delayed_jobs_priority", using: :btree
+
   create_table "feed_preferences", force: true do |t|
     t.integer  "user_id"
-    t.integer  "feed_id"
     t.datetime "last_visited"
     t.datetime "previous_last_visited"
     t.integer  "selected_range"
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.text     "feed_uid"
   end
 
   create_table "feeds", force: true do |t|
@@ -92,6 +131,7 @@ ActiveRecord::Schema.define(version: 20140305044305) do
   end
 
   add_index "feeds", ["source"], name: "index_feeds_on_source", using: :btree
+  add_index "feeds", ["uid", "last_paper_date"], name: "index_feeds_on_uid_and_last_paper_date", order: {"uid"=>:desc, "last_paper_date"=>:desc}, using: :btree
   add_index "feeds", ["uid"], name: "index_feeds_on_uid", unique: true, using: :btree
 
   create_table "papers", force: true do |t|
@@ -116,6 +156,7 @@ ActiveRecord::Schema.define(version: 20140305044305) do
     t.integer  "comments_count",  default: 0, null: false
     t.datetime "pubdate"
     t.text     "author_str",                  null: false
+    t.integer  "versions_count",  default: 1, null: false
   end
 
   add_index "papers", ["abs_url"], name: "index_papers_on_abs_url", unique: true, using: :btree
@@ -138,6 +179,16 @@ ActiveRecord::Schema.define(version: 20140305044305) do
   add_index "scites", ["paper_uid"], name: "index_scites_on_paper_uid", using: :btree
   add_index "scites", ["user_id"], name: "index_scites_on_user_id", using: :btree
 
+  create_table "sessions", force: true do |t|
+    t.string   "session_id", null: false
+    t.text     "data"
+    t.datetime "created_at"
+    t.datetime "updated_at"
+  end
+
+  add_index "sessions", ["session_id"], name: "index_sessions_on_session_id", unique: true, using: :btree
+  add_index "sessions", ["updated_at"], name: "index_sessions_on_updated_at", using: :btree
+
   create_table "subscriptions", force: true do |t|
     t.integer  "user_id"
     t.datetime "created_at",              null: false
@@ -149,24 +200,39 @@ ActiveRecord::Schema.define(version: 20140305044305) do
   add_index "subscriptions", ["feed_uid"], name: "index_subscriptions_on_feed_uid", using: :btree
   add_index "subscriptions", ["user_id"], name: "index_subscriptions_on_user_id", using: :btree
 
+  create_table "system", force: true do |t|
+    t.text     "alert",      default: "", null: false
+    t.datetime "created_at"
+    t.datetime "updated_at"
+  end
+
   create_table "users", force: true do |t|
     t.text     "fullname"
     t.text     "email"
     t.text     "remember_token"
-    t.datetime "created_at",                              null: false
-    t.datetime "updated_at",                              null: false
+    t.datetime "created_at",                                        null: false
+    t.datetime "updated_at",                                        null: false
     t.text     "password_digest"
-    t.integer  "scites_count",           default: 0
+    t.integer  "scites_count",                     default: 0
     t.text     "password_reset_token"
     t.datetime "password_reset_sent_at"
     t.text     "confirmation_token"
-    t.boolean  "active",                 default: false
-    t.integer  "comments_count",         default: 0
+    t.boolean  "active",                           default: false
+    t.integer  "comments_count",                   default: 0
     t.datetime "confirmation_sent_at"
-    t.integer  "subscriptions_count",    default: 0
-    t.boolean  "expand_abstracts",       default: false
-    t.text     "account_status",         default: "user"
-    t.text     "username"
+    t.integer  "subscriptions_count",              default: 0
+    t.boolean  "expand_abstracts",                 default: false
+    t.text     "account_status",                   default: "user"
+    t.text     "username",                                          null: false
+    t.text     "organization",                     default: "",     null: false
+    t.text     "about",                            default: "",     null: false
+    t.text     "url",                              default: "",     null: false
+    t.text     "location",                         default: "",     null: false
+    t.text     "author_identifier",                default: "",     null: false
+    t.integer  "papers_count",                     default: 0,      null: false
+    t.boolean  "email_about_replies",              default: true
+    t.boolean  "email_about_comments_on_authored", default: true
+    t.boolean  "email_about_comments_on_scited",   default: false
   end
 
   add_index "users", ["email"], name: "index_users_on_email", unique: true, using: :btree
