@@ -1,156 +1,176 @@
 Feed.create!([
-  {uid: "astro-ph", source: "arxiv", fullname: "Astrophysics", position: 0, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "astro-ph.CO", source: "arxiv", fullname: "Cosmology and Extragalactic Astrophysics", position: 1, subscriptions_count: 0, last_paper_date: nil, parent_uid: "astro-ph"},
-  {uid: "astro-ph.EP", source: "arxiv", fullname: "Earth and Planetary Astrophysics", position: 2, subscriptions_count: 0, last_paper_date: nil, parent_uid: "astro-ph"},
-  {uid: "astro-ph.GA", source: "arxiv", fullname: "Galaxy Astrophysics", position: 3, subscriptions_count: 0, last_paper_date: nil, parent_uid: "astro-ph"},
-  {uid: "astro-ph.HE", source: "arxiv", fullname: "High Energy Astrophysical Phenomena", position: 4, subscriptions_count: 0, last_paper_date: nil, parent_uid: "astro-ph"},
-  {uid: "astro-ph.IM", source: "arxiv", fullname: "Instrumentation and Methods for Astrophysics", position: 5, subscriptions_count: 0, last_paper_date: nil, parent_uid: "astro-ph"},
-  {uid: "astro-ph.SR", source: "arxiv", fullname: "Solar and Stellar Astrophysics", position: 6, subscriptions_count: 0, last_paper_date: nil, parent_uid: "astro-ph"},
-  {uid: "cond-mat", source: "arxiv", fullname: "Condensed Matter", position: 7, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "cond-mat.dis-nn", source: "arxiv", fullname: "Disordered Systems and Neural Networks", position: 8, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cond-mat"},
-  {uid: "cond-mat.mtrl-sci", source: "arxiv", fullname: "Materials Science", position: 9, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cond-mat"},
-  {uid: "cond-mat.mes-hall", source: "arxiv", fullname: "Mesoscale and Nanoscale Physics", position: 10, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cond-mat"},
-  {uid: "cond-mat.other", source: "arxiv", fullname: "Other Condensed Matter", position: 11, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cond-mat"},
-  {uid: "cond-mat.quant-gas", source: "arxiv", fullname: "Quantum Gases", position: 12, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cond-mat"},
-  {uid: "cond-mat.soft", source: "arxiv", fullname: "Soft Condensed Matter", position: 13, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cond-mat"},
-  {uid: "cond-mat.stat-mech", source: "arxiv", fullname: "Statistical Mechanics", position: 14, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cond-mat"},
-  {uid: "cond-mat.str-el", source: "arxiv", fullname: "Strongly Correlated Electrons", position: 15, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cond-mat"},
-  {uid: "cond-mat.supr-con", source: "arxiv", fullname: "Superconductivity", position: 16, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cond-mat"},
-  {uid: "gr-qc", source: "arxiv", fullname: "General Relativity and Quantum Cosmology", position: 17, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "hep-ex", source: "arxiv", fullname: "High Energy Physics - Experiment", position: 18, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "hep-lat", source: "arxiv", fullname: "High Energy Physics - Lattice", position: 19, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "hep-ph", source: "arxiv", fullname: "High Energy Physics - Phenomenology", position: 20, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "hep-th", source: "arxiv", fullname: "High Energy Physics - Theory", position: 21, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "math-ph", source: "arxiv", fullname: "Mathematical Physics", position: 22, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "nlin", source: "arxiv", fullname: "Nonlinear Sciences", position: 23, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "nlin.AO", source: "arxiv", fullname: "Adaptation and Self-Organizing Systems", position: 24, subscriptions_count: 0, last_paper_date: nil, parent_uid: "nlin"},
-  {uid: "nlin.CG", source: "arxiv", fullname: "Cellular Automata and Lattice Gases", position: 25, subscriptions_count: 0, last_paper_date: nil, parent_uid: "nlin"},
-  {uid: "nlin.CD", source: "arxiv", fullname: "Chaotic Dynamics", position: 26, subscriptions_count: 0, last_paper_date: nil, parent_uid: "nlin"},
-  {uid: "nlin.SI", source: "arxiv", fullname: "Exactly Solvable and Integrable Systems", position: 27, subscriptions_count: 0, last_paper_date: nil, parent_uid: "nlin"},
-  {uid: "nlin.PS", source: "arxiv", fullname: "Pattern Formation and Solitons", position: 28, subscriptions_count: 0, last_paper_date: nil, parent_uid: "nlin"},
-  {uid: "nucl-ex", source: "arxiv", fullname: "Nuclear Experiment", position: 29, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "nucl-th", source: "arxiv", fullname: "Nuclear Theory", position: 30, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "physics", source: "arxiv", fullname: "More Physics", position: 31, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "physics.acc-ph", source: "arxiv", fullname: "Accelerator Physics", position: 32, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.ao-ph", source: "arxiv", fullname: "Atmospheric and Oceanic Physics", position: 33, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.atom-ph", source: "arxiv", fullname: "Atomic Physics", position: 34, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.atm-clus", source: "arxiv", fullname: "Atomic and Molecular Clusters", position: 35, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.bio-ph", source: "arxiv", fullname: "Biological Physics", position: 36, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.chem-ph", source: "arxiv", fullname: "Chemical Physics", position: 37, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.class-ph", source: "arxiv", fullname: "Classical Physics", position: 38, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.comp-ph", source: "arxiv", fullname: "Computational Physics", position: 39, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.data-an", source: "arxiv", fullname: "Data Analysis, Statistics and Probability", position: 40, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.flu-dyn", source: "arxiv", fullname: "Fluid Dynamics", position: 41, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.gen-ph", source: "arxiv", fullname: "General Physics", position: 42, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.geo-ph", source: "arxiv", fullname: "Geophysics", position: 43, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.hist-ph", source: "arxiv", fullname: "History and Philosophy of Physics", position: 44, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.ins-det", source: "arxiv", fullname: "Instrumentation and Detectors", position: 45, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.med-ph", source: "arxiv", fullname: "Medical Physics", position: 46, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.optics", source: "arxiv", fullname: "Optics", position: 47, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.ed-ph", source: "arxiv", fullname: "Physics Education", position: 48, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.soc-ph", source: "arxiv", fullname: "Physics and Society", position: 49, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.plasm-ph", source: "arxiv", fullname: "Plasma Physics", position: 50, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.pop-ph", source: "arxiv", fullname: "Popular Physics", position: 51, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "physics.space-ph", source: "arxiv", fullname: "Space Physics", position: 52, subscriptions_count: 0, last_paper_date: nil, parent_uid: "physics"},
-  {uid: "quant-ph", source: "arxiv", fullname: "Quantum Physics", position: 53, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "math", source: "arxiv", fullname: "Mathematics", position: 54, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "math.AG", source: "arxiv", fullname: "Algebraic Geometry", position: 55, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.AT", source: "arxiv", fullname: "Algebraic Topology", position: 56, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.AP", source: "arxiv", fullname: "Analysis of PDEs", position: 57, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.CT", source: "arxiv", fullname: "Category Theory", position: 58, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.CA", source: "arxiv", fullname: "Classical Analysis and ODEs", position: 59, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.CO", source: "arxiv", fullname: "Combinatorics", position: 60, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.AC", source: "arxiv", fullname: "Commutative Algebra", position: 61, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.CV", source: "arxiv", fullname: "Complex Variables", position: 62, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.DG", source: "arxiv", fullname: "Differential Geometry", position: 63, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.DS", source: "arxiv", fullname: "Dynamical Systems", position: 64, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.FA", source: "arxiv", fullname: "Functional Analysis", position: 65, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.GM", source: "arxiv", fullname: "General Mathematics", position: 66, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.GN", source: "arxiv", fullname: "General Topology", position: 67, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.GT", source: "arxiv", fullname: "Geometric Topology", position: 68, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.GR", source: "arxiv", fullname: "Group Theory", position: 69, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.HO", source: "arxiv", fullname: "History and Overview", position: 70, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.IT", source: "arxiv", fullname: "Information Theory", position: 71, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.KT", source: "arxiv", fullname: "K-Theory and Homology", position: 72, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.LO", source: "arxiv", fullname: "Logic", position: 73, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.MP", source: "arxiv", fullname: "Mathematical Physics", position: 74, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.MG", source: "arxiv", fullname: "Metric Geometry", position: 75, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.NT", source: "arxiv", fullname: "Number Theory", position: 76, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.NA", source: "arxiv", fullname: "Numerical Analysis", position: 77, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.OA", source: "arxiv", fullname: "Operator Algebras", position: 78, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.OC", source: "arxiv", fullname: "Optimization and Control", position: 79, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.PR", source: "arxiv", fullname: "Probability", position: 80, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.QA", source: "arxiv", fullname: "Quantum Algebra", position: 81, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.RT", source: "arxiv", fullname: "Representation Theory", position: 82, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.RA", source: "arxiv", fullname: "Rings and Algebras", position: 83, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.SP", source: "arxiv", fullname: "Spectral Theory", position: 84, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.ST", source: "arxiv", fullname: "Statistics Theory", position: 85, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "math.SG", source: "arxiv", fullname: "Symplectic Geometry", position: 86, subscriptions_count: 0, last_paper_date: nil, parent_uid: "math"},
-  {uid: "cs", source: "arxiv", fullname: "Computer Science", position: 87, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "cs.AI", source: "arxiv", fullname: "Artificial Intelligence", position: 88, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.CL", source: "arxiv", fullname: "Computation and Language", position: 89, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.CC", source: "arxiv", fullname: "Computational Complexity", position: 90, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.CE", source: "arxiv", fullname: "Computational Engineering, Finance, and Science", position: 91, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.CG", source: "arxiv", fullname: "Computational Geometry", position: 92, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.GT", source: "arxiv", fullname: "Computer Science and Game Theory", position: 93, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.CV", source: "arxiv", fullname: "Computer Vision and Pattern Recognition", position: 94, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.CY", source: "arxiv", fullname: "Computers and Society", position: 95, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.CR", source: "arxiv", fullname: "Cryptography and Security", position: 96, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.DS", source: "arxiv", fullname: "Data Structures and Algorithms", position: 97, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.DB", source: "arxiv", fullname: "Databases", position: 98, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.DL", source: "arxiv", fullname: "Digital Libraries", position: 99, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.DM", source: "arxiv", fullname: "Discrete Mathematics", position: 100, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.DC", source: "arxiv", fullname: "Distributed, Parallel, and Cluster Computing", position: 101, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.ET", source: "arxiv", fullname: "Emerging Technologies", position: 102, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.FL", source: "arxiv", fullname: "Formal Languages and Automata Theory", position: 103, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.GL", source: "arxiv", fullname: "General Literature", position: 104, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.GR", source: "arxiv", fullname: "Graphics", position: 105, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.AR", source: "arxiv", fullname: "Hardware Architecture", position: 106, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.HC", source: "arxiv", fullname: "Human-Computer Interaction", position: 107, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.IR", source: "arxiv", fullname: "Information Retrieval", position: 108, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.IT", source: "arxiv", fullname: "Information Theory", position: 109, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.LG", source: "arxiv", fullname: "Learning", position: 110, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.LO", source: "arxiv", fullname: "Logic in Computer Science", position: 111, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.MS", source: "arxiv", fullname: "Mathematical Software", position: 112, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.MA", source: "arxiv", fullname: "Multiagent Systems", position: 113, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.MM", source: "arxiv", fullname: "Multimedia", position: 114, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.NI", source: "arxiv", fullname: "Networking and Internet Architecture", position: 115, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.NE", source: "arxiv", fullname: "Neural and Evolutionary Computing", position: 116, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.NA", source: "arxiv", fullname: "Numerical Analysis", position: 117, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.OS", source: "arxiv", fullname: "Operating Systems", position: 118, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.OH", source: "arxiv", fullname: "Other Computer Science", position: 119, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.PF", source: "arxiv", fullname: "Performance", position: 120, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.PL", source: "arxiv", fullname: "Programming Languages", position: 121, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.RO", source: "arxiv", fullname: "Robotics", position: 122, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.SI", source: "arxiv", fullname: "Social and Information Networks", position: 123, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.SE", source: "arxiv", fullname: "Software Engineering", position: 124, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.SD", source: "arxiv", fullname: "Sound", position: 125, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.SC", source: "arxiv", fullname: "Symbolic Computation", position: 126, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "cs.SY", source: "arxiv", fullname: "Systems and Control", position: 127, subscriptions_count: 0, last_paper_date: nil, parent_uid: "cs"},
-  {uid: "q-bio", source: "arxiv", fullname: "Quantitative Biology", position: 128, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "q-bio.BM", source: "arxiv", fullname: "Biomolecules", position: 129, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-bio"},
-  {uid: "q-bio.CB", source: "arxiv", fullname: "Cell Behavior", position: 130, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-bio"},
-  {uid: "q-bio.GN", source: "arxiv", fullname: "Genomics", position: 131, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-bio"},
-  {uid: "q-bio.MN", source: "arxiv", fullname: "Molecular Networks", position: 132, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-bio"},
-  {uid: "q-bio.NC", source: "arxiv", fullname: "Neurons and Cognition", position: 133, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-bio"},
-  {uid: "q-bio.OT", source: "arxiv", fullname: "Other Quantitative Biology", position: 134, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-bio"},
-  {uid: "q-bio.PE", source: "arxiv", fullname: "Populations and Evolution", position: 135, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-bio"},
-  {uid: "q-bio.QM", source: "arxiv", fullname: "Quantitative Methods", position: 136, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-bio"},
-  {uid: "q-bio.SC", source: "arxiv", fullname: "Subcellular Processes", position: 137, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-bio"},
-  {uid: "q-bio.TO", source: "arxiv", fullname: "Tissues and Organs", position: 138, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-bio"},
-  {uid: "q-fin", source: "arxiv", fullname: "Quantitative Finance", position: 139, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "q-fin.CP", source: "arxiv", fullname: "Computational Finance", position: 140, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-fin"},
-  {uid: "q-fin.GN", source: "arxiv", fullname: "General Finance", position: 141, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-fin"},
-  {uid: "q-fin.PM", source: "arxiv", fullname: "Portfolio Management", position: 142, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-fin"},
-  {uid: "q-fin.PR", source: "arxiv", fullname: "Pricing of Securities", position: 143, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-fin"},
-  {uid: "q-fin.RM", source: "arxiv", fullname: "Risk Management", position: 144, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-fin"},
-  {uid: "q-fin.ST", source: "arxiv", fullname: "Statistical Finance", position: 145, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-fin"},
-  {uid: "q-fin.TR", source: "arxiv", fullname: "Trading and Market Microstructure", position: 146, subscriptions_count: 0, last_paper_date: nil, parent_uid: "q-fin"},
-  {uid: "stat", source: "arxiv", fullname: "Statistics", position: 147, subscriptions_count: 0, last_paper_date: nil, parent_uid: nil},
-  {uid: "stat.AP", source: "arxiv", fullname: "Applications", position: 148, subscriptions_count: 0, last_paper_date: nil, parent_uid: "stat"},
-  {uid: "stat.CO", source: "arxiv", fullname: "Computation", position: 149, subscriptions_count: 0, last_paper_date: nil, parent_uid: "stat"},
-  {uid: "stat.ML", source: "arxiv", fullname: "Machine Learning", position: 150, subscriptions_count: 0, last_paper_date: nil, parent_uid: "stat"},
-  {uid: "stat.ME", source: "arxiv", fullname: "Methodology", position: 151, subscriptions_count: 0, last_paper_date: nil, parent_uid: "stat"},
-  {uid: "stat.OT", source: "arxiv", fullname: "Other Statistics", position: 152, subscriptions_count: 0, last_paper_date: nil, parent_uid: "stat"},
-  {uid: "stat.TH", source: "arxiv", fullname: "Statistics Theory", position: 153, subscriptions_count: 0, last_paper_date: nil, parent_uid: "stat"}
+  {uid: "alg-geom", source: "arxiv", fullname: "alg-geom", position: 0, subscriptions_count: 0, last_paper_date: "1998-01-02 01:00:00", parent_uid: nil},
+  {uid: "ao-sci", source: "arxiv", fullname: "ao-sci", position: 0, subscriptions_count: 0, last_paper_date: "1996-09-03 01:00:00", parent_uid: nil},
+  {uid: "dg-ga", source: "arxiv", fullname: "dg-ga", position: 0, subscriptions_count: 0, last_paper_date: "1997-12-31 01:00:00", parent_uid: nil},
+  {uid: "chao-dyn", source: "arxiv", fullname: "chao-dyn", position: 0, subscriptions_count: 0, last_paper_date: "1999-12-30 01:00:00", parent_uid: nil},
+  {uid: "acc-phys", source: "arxiv", fullname: "acc-phys", position: 0, subscriptions_count: 0, last_paper_date: "1996-09-25 01:00:00", parent_uid: nil},
+  {uid: "q-alg", source: "arxiv", fullname: "q-alg", position: 0, subscriptions_count: 0, last_paper_date: "1997-12-31 01:00:00", parent_uid: nil},
+  {uid: "atom-ph", source: "arxiv", fullname: "atom-ph", position: 0, subscriptions_count: 0, last_paper_date: "1996-10-01 01:00:00", parent_uid: nil},
+  {uid: "bayes-an", source: "arxiv", fullname: "bayes-an", position: 0, subscriptions_count: 0, last_paper_date: "1996-11-22 01:00:00", parent_uid: nil},
+  {uid: "supr-con", source: "arxiv", fullname: "supr-con", position: 0, subscriptions_count: 0, last_paper_date: "1996-09-26 01:00:00", parent_uid: nil},
+  {uid: "chem-ph", source: "arxiv", fullname: "chem-ph", position: 0, subscriptions_count: 0, last_paper_date: "1996-10-01 01:00:00", parent_uid: nil},
+  {uid: "funct-an", source: "arxiv", fullname: "funct-an", position: 0, subscriptions_count: 0, last_paper_date: "1997-12-26 01:00:00", parent_uid: nil},
+  {uid: "physics.pop-ph", source: "arxiv", fullname: "Popular Physics", position: 51, subscriptions_count: 11, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "mtrl-th", source: "arxiv", fullname: "mtrl-th", position: 0, subscriptions_count: 0, last_paper_date: "1996-09-27 01:00:00", parent_uid: nil},
+  {uid: "cs", source: "arxiv", fullname: "Computer Science", position: 87, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "plasm-ph", source: "arxiv", fullname: "plasm-ph", position: 0, subscriptions_count: 0, last_paper_date: "1996-09-09 01:00:00", parent_uid: nil},
+  {uid: "adap-org", source: "arxiv", fullname: "adap-org", position: 0, subscriptions_count: 0, last_paper_date: "1999-12-30 01:00:00", parent_uid: nil},
+  {uid: "comp-gas", source: "arxiv", fullname: "comp-gas", position: 0, subscriptions_count: 0, last_paper_date: "1999-11-23 01:00:00", parent_uid: nil},
+  {uid: "solv-int", source: "arxiv", fullname: "solv-int", position: 0, subscriptions_count: 0, last_paper_date: "1999-12-30 01:00:00", parent_uid: nil},
+  {uid: "cs.GL", source: "arxiv", fullname: "General Literature", position: 104, subscriptions_count: 7, last_paper_date: "2014-02-06 01:00:00", parent_uid: "cs"},
+  {uid: "cs.GR", source: "arxiv", fullname: "Graphics", position: 105, subscriptions_count: 5, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "patt-sol", source: "arxiv", fullname: "patt-sol", position: 0, subscriptions_count: 0, last_paper_date: "1999-12-24 01:00:00", parent_uid: nil},
+  {uid: "cmp-lg", source: "arxiv", fullname: "cmp-lg", position: 0, subscriptions_count: 0, last_paper_date: "1998-09-15 01:00:00", parent_uid: nil},
+  {uid: "physics.acc-ph", source: "arxiv", fullname: "Accelerator Physics", position: 32, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.gen-ph", source: "arxiv", fullname: "General Physics", position: 42, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.space-ph", source: "arxiv", fullname: "Space Physics", position: 52, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "math.GT", source: "arxiv", fullname: "Geometric Topology", position: 68, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.GR", source: "arxiv", fullname: "Group Theory", position: 69, subscriptions_count: 14, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.KT", source: "arxiv", fullname: "K-Theory and Homology", position: 72, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.LO", source: "arxiv", fullname: "Logic", position: 73, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.MP", source: "arxiv", fullname: "Mathematical Physics", position: 74, subscriptions_count: 29, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.MG", source: "arxiv", fullname: "Metric Geometry", position: 75, subscriptions_count: 11, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.NT", source: "arxiv", fullname: "Number Theory", position: 76, subscriptions_count: 12, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.NA", source: "arxiv", fullname: "Numerical Analysis", position: 77, subscriptions_count: 9, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.OA", source: "arxiv", fullname: "Operator Algebras", position: 78, subscriptions_count: 17, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.OC", source: "arxiv", fullname: "Optimization and Control", position: 79, subscriptions_count: 9, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.PR", source: "arxiv", fullname: "Probability", position: 80, subscriptions_count: 16, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "cs.HC", source: "arxiv", fullname: "Human-Computer Interaction", position: 107, subscriptions_count: 4, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "math.RT", source: "arxiv", fullname: "Representation Theory", position: 82, subscriptions_count: 21, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.RA", source: "arxiv", fullname: "Rings and Algebras", position: 83, subscriptions_count: 11, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.SP", source: "arxiv", fullname: "Spectral Theory", position: 84, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.ST", source: "arxiv", fullname: "Statistics Theory", position: 85, subscriptions_count: 11, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.SG", source: "arxiv", fullname: "Symplectic Geometry", position: 86, subscriptions_count: 9, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "stat", source: "arxiv", fullname: "Statistics", position: 149, subscriptions_count: 2, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "stat.AP", source: "arxiv", fullname: "Applications", position: 150, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "stat"},
+  {uid: "stat.CO", source: "arxiv", fullname: "Computation", position: 151, subscriptions_count: 13, last_paper_date: "2014-05-12 01:00:00", parent_uid: "stat"},
+  {uid: "stat.ML", source: "arxiv", fullname: "Machine Learning", position: 152, subscriptions_count: 15, last_paper_date: "2014-05-12 01:00:00", parent_uid: "stat"},
+  {uid: "stat.ME", source: "arxiv", fullname: "Methodology", position: 153, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "stat"},
+  {uid: "stat.OT", source: "arxiv", fullname: "Other Statistics", position: 154, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "stat"},
+  {uid: "stat.TH", source: "arxiv", fullname: "Statistics Theory", position: 155, subscriptions_count: 9, last_paper_date: "2014-05-12 01:00:00", parent_uid: "stat"},
+  {uid: "cs.AI", source: "arxiv", fullname: "Artificial Intelligence", position: 88, subscriptions_count: 18, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.CC", source: "arxiv", fullname: "Computational Complexity", position: 90, subscriptions_count: 48, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.CE", source: "arxiv", fullname: "Computational Engineering, Finance, and Science", position: 91, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.CL", source: "arxiv", fullname: "Computation and Language", position: 89, subscriptions_count: 13, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.CY", source: "arxiv", fullname: "Computers and Society", position: 95, subscriptions_count: 5, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.CV", source: "arxiv", fullname: "Computer Vision and Pattern Recognition", position: 94, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.CR", source: "arxiv", fullname: "Cryptography and Security", position: 96, subscriptions_count: 13, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.IR", source: "arxiv", fullname: "Information Retrieval", position: 108, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.DS", source: "arxiv", fullname: "Data Structures and Algorithms", position: 97, subscriptions_count: 21, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.DL", source: "arxiv", fullname: "Digital Libraries", position: 99, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.DM", source: "arxiv", fullname: "Discrete Mathematics", position: 100, subscriptions_count: 24, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.DC", source: "arxiv", fullname: "Distributed, Parallel, and Cluster Computing", position: 101, subscriptions_count: 14, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.FL", source: "arxiv", fullname: "Formal Languages and Automata Theory", position: 103, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.IT", source: "arxiv", fullname: "Information Theory", position: 109, subscriptions_count: 37, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.LG", source: "arxiv", fullname: "Learning", position: 110, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.LO", source: "arxiv", fullname: "Logic in Computer Science", position: 111, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.MS", source: "arxiv", fullname: "Mathematical Software", position: 112, subscriptions_count: 13, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.MM", source: "arxiv", fullname: "Multimedia", position: 114, subscriptions_count: 5, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "q-bio.TO", source: "arxiv", fullname: "Tissues and Organs", position: 138, subscriptions_count: 3, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-bio"},
+  {uid: "q-fin.CP", source: "arxiv", fullname: "Computational Finance", position: 140, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-fin"},
+  {uid: "physics.hist-ph", source: "arxiv", fullname: "History and Philosophy of Physics", position: 44, subscriptions_count: 16, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "q-fin.EC", source: "arxiv", fullname: "Economics", position: 141, subscriptions_count: 2, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-fin"},
+  {uid: "q-fin.GN", source: "arxiv", fullname: "General Finance", position: 142, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-fin"},
+  {uid: "q-fin.PR", source: "arxiv", fullname: "Pricing of Securities", position: 145, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-fin"},
+  {uid: "q-fin.RM", source: "arxiv", fullname: "Risk Management", position: 146, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-fin"},
+  {uid: "physics.ao-ph", source: "arxiv", fullname: "Atmospheric and Oceanic Physics", position: 33, subscriptions_count: 4, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "math.HO", source: "arxiv", fullname: "History and Overview", position: 70, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.QA", source: "arxiv", fullname: "Quantum Algebra", position: 81, subscriptions_count: 19, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "cs.CG", source: "arxiv", fullname: "Computational Geometry", position: 92, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.GT", source: "arxiv", fullname: "Computer Science and Game Theory", position: 93, subscriptions_count: 12, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.DB", source: "arxiv", fullname: "Databases", position: 98, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.ET", source: "arxiv", fullname: "Emerging Technologies", position: 102, subscriptions_count: 9, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.AR", source: "arxiv", fullname: "Hardware Architecture", position: 106, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.MA", source: "arxiv", fullname: "Multiagent Systems", position: 113, subscriptions_count: 3, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.NI", source: "arxiv", fullname: "Networking and Internet Architecture", position: 115, subscriptions_count: 5, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.NE", source: "arxiv", fullname: "Neural and Evolutionary Computing", position: 116, subscriptions_count: 4, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.NA", source: "arxiv", fullname: "Numerical Analysis", position: 117, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.OS", source: "arxiv", fullname: "Operating Systems", position: 118, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.OH", source: "arxiv", fullname: "Other Computer Science", position: 119, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.PF", source: "arxiv", fullname: "Performance", position: 120, subscriptions_count: 3, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.PL", source: "arxiv", fullname: "Programming Languages", position: 121, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.RO", source: "arxiv", fullname: "Robotics", position: 122, subscriptions_count: 4, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.SI", source: "arxiv", fullname: "Social and Information Networks", position: 123, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.SE", source: "arxiv", fullname: "Software Engineering", position: 124, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.SD", source: "arxiv", fullname: "Sound", position: 125, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.SC", source: "arxiv", fullname: "Symbolic Computation", position: 126, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "cs.SY", source: "arxiv", fullname: "Systems and Control", position: 127, subscriptions_count: 3, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cs"},
+  {uid: "q-bio", source: "arxiv", fullname: "Quantitative Biology", position: 128, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "q-bio.BM", source: "arxiv", fullname: "Biomolecules", position: 129, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-bio"},
+  {uid: "q-bio.CB", source: "arxiv", fullname: "Cell Behavior", position: 130, subscriptions_count: 4, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-bio"},
+  {uid: "q-bio.GN", source: "arxiv", fullname: "Genomics", position: 131, subscriptions_count: 4, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-bio"},
+  {uid: "q-bio.MN", source: "arxiv", fullname: "Molecular Networks", position: 132, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-bio"},
+  {uid: "cond-mat", source: "arxiv", fullname: "Condensed Matter", position: 7, subscriptions_count: 25, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "q-bio.NC", source: "arxiv", fullname: "Neurons and Cognition", position: 133, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-bio"},
+  {uid: "q-bio.OT", source: "arxiv", fullname: "Other Quantitative Biology", position: 134, subscriptions_count: 3, last_paper_date: "2014-04-16 01:00:00", parent_uid: "q-bio"},
+  {uid: "q-bio.PE", source: "arxiv", fullname: "Populations and Evolution", position: 135, subscriptions_count: 5, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-bio"},
+  {uid: "q-bio.QM", source: "arxiv", fullname: "Quantitative Methods", position: 136, subscriptions_count: 5, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-bio"},
+  {uid: "q-bio.SC", source: "arxiv", fullname: "Subcellular Processes", position: 137, subscriptions_count: 5, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-bio"},
+  {uid: "q-fin", source: "arxiv", fullname: "Quantitative Finance", position: 139, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "q-fin.MF", source: "arxiv", fullname: "Mathematical Finance", position: 143, subscriptions_count: 2, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-fin"},
+  {uid: "q-fin.PM", source: "arxiv", fullname: "Portfolio Management", position: 144, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-fin"},
+  {uid: "q-fin.ST", source: "arxiv", fullname: "Statistical Finance", position: 147, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-fin"},
+  {uid: "q-fin.TR", source: "arxiv", fullname: "Trading and Market Microstructure", position: 148, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "q-fin"},
+  {uid: "nlin", source: "arxiv", fullname: "Nonlinear Sciences", position: 23, subscriptions_count: 4, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "nlin.AO", source: "arxiv", fullname: "Adaptation and Self-Organizing Systems", position: 24, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "nlin"},
+  {uid: "nlin.CG", source: "arxiv", fullname: "Cellular Automata and Lattice Gases", position: 25, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "nlin"},
+  {uid: "nlin.CD", source: "arxiv", fullname: "Chaotic Dynamics", position: 26, subscriptions_count: 11, last_paper_date: "2014-05-12 01:00:00", parent_uid: "nlin"},
+  {uid: "nlin.SI", source: "arxiv", fullname: "Exactly Solvable and Integrable Systems", position: 27, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "nlin"},
+  {uid: "nlin.PS", source: "arxiv", fullname: "Pattern Formation and Solitons", position: 28, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "nlin"},
+  {uid: "physics.geo-ph", source: "arxiv", fullname: "Geophysics", position: 43, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "hep-th", source: "arxiv", fullname: "High Energy Physics - Theory", position: 21, subscriptions_count: 42, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "quant-ph", source: "arxiv", fullname: "Quantum Physics", position: 53, subscriptions_count: 939, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "nucl-th", source: "arxiv", fullname: "Nuclear Theory", position: 30, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "math-ph", source: "arxiv", fullname: "Mathematical Physics", position: 22, subscriptions_count: 47, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "physics.optics", source: "arxiv", fullname: "Optics", position: 47, subscriptions_count: 15, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "hep-ph", source: "arxiv", fullname: "High Energy Physics - Phenomenology", position: 20, subscriptions_count: 11, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "hep-ex", source: "arxiv", fullname: "High Energy Physics - Experiment", position: 18, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "physics.plasm-ph", source: "arxiv", fullname: "Plasma Physics", position: 50, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "astro-ph.EP", source: "arxiv", fullname: "Earth and Planetary Astrophysics", position: 3, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "astro-ph"},
+  {uid: "astro-ph.IM", source: "arxiv", fullname: "Instrumentation and Methods for Astrophysics", position: 5, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "astro-ph"},
+  {uid: "math", source: "arxiv", fullname: "Mathematics", position: 54, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "astro-ph", source: "arxiv", fullname: "Astrophysics", position: 0, subscriptions_count: 11, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "astro-ph.CO", source: "arxiv", fullname: "Cosmology and Nongalactic Astrophysics", position: 2, subscriptions_count: 12, last_paper_date: "2014-05-12 01:00:00", parent_uid: "astro-ph"},
+  {uid: "math.AG", source: "arxiv", fullname: "Algebraic Geometry", position: 55, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "astro-ph.HE", source: "arxiv", fullname: "High Energy Astrophysical Phenomena", position: 4, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "astro-ph"},
+  {uid: "astro-ph.SR", source: "arxiv", fullname: "Solar and Stellar Astrophysics", position: 6, subscriptions_count: 4, last_paper_date: "2014-05-12 01:00:00", parent_uid: "astro-ph"},
+  {uid: "cond-mat.mtrl-sci", source: "arxiv", fullname: "Materials Science", position: 9, subscriptions_count: 14, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cond-mat"},
+  {uid: "cond-mat.str-el", source: "arxiv", fullname: "Strongly Correlated Electrons", position: 15, subscriptions_count: 34, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cond-mat"},
+  {uid: "cond-mat.supr-con", source: "arxiv", fullname: "Superconductivity", position: 16, subscriptions_count: 13, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cond-mat"},
+  {uid: "cond-mat.stat-mech", source: "arxiv", fullname: "Statistical Mechanics", position: 14, subscriptions_count: 40, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cond-mat"},
+  {uid: "cond-mat.mes-hall", source: "arxiv", fullname: "Mesoscale and Nanoscale Physics", position: 10, subscriptions_count: 18, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cond-mat"},
+  {uid: "gr-qc", source: "arxiv", fullname: "General Relativity and Quantum Cosmology", position: 17, subscriptions_count: 43, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "astro-ph.GA", source: "arxiv", fullname: "Astrophysics of Galaxies", position: 1, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "astro-ph"},
+  {uid: "physics", source: "arxiv", fullname: "More Physics", position: 31, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "physics.atom-ph", source: "arxiv", fullname: "Atomic Physics", position: 34, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "math.AT", source: "arxiv", fullname: "Algebraic Topology", position: 56, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.AP", source: "arxiv", fullname: "Analysis of PDEs", position: 57, subscriptions_count: 12, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.CA", source: "arxiv", fullname: "Classical Analysis and ODEs", position: 59, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.CO", source: "arxiv", fullname: "Combinatorics", position: 60, subscriptions_count: 28, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.CV", source: "arxiv", fullname: "Complex Variables", position: 62, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.DG", source: "arxiv", fullname: "Differential Geometry", position: 63, subscriptions_count: 12, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.DS", source: "arxiv", fullname: "Dynamical Systems", position: 64, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.IT", source: "arxiv", fullname: "Information Theory", position: 71, subscriptions_count: 22, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "physics.soc-ph", source: "arxiv", fullname: "Physics and Society", position: 49, subscriptions_count: 8, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.ed-ph", source: "arxiv", fullname: "Physics Education", position: 48, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "nucl-ex", source: "arxiv", fullname: "Nuclear Experiment", position: 29, subscriptions_count: 5, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "cond-mat.other", source: "arxiv", fullname: "Other Condensed Matter", position: 11, subscriptions_count: 12, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cond-mat"},
+  {uid: "cond-mat.soft", source: "arxiv", fullname: "Soft Condensed Matter", position: 13, subscriptions_count: 10, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cond-mat"},
+  {uid: "cond-mat.quant-gas", source: "arxiv", fullname: "Quantum Gases", position: 12, subscriptions_count: 17, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cond-mat"},
+  {uid: "cond-mat.dis-nn", source: "arxiv", fullname: "Disordered Systems and Neural Networks", position: 8, subscriptions_count: 17, last_paper_date: "2014-05-12 01:00:00", parent_uid: "cond-mat"},
+  {uid: "hep-lat", source: "arxiv", fullname: "High Energy Physics - Lattice", position: 19, subscriptions_count: 17, last_paper_date: "2014-05-12 01:00:00", parent_uid: nil},
+  {uid: "physics.flu-dyn", source: "arxiv", fullname: "Fluid Dynamics", position: 41, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.ins-det", source: "arxiv", fullname: "Instrumentation and Detectors", position: 45, subscriptions_count: 3, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.med-ph", source: "arxiv", fullname: "Medical Physics", position: 46, subscriptions_count: 4, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.atm-clus", source: "arxiv", fullname: "Atomic and Molecular Clusters", position: 35, subscriptions_count: 4, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.bio-ph", source: "arxiv", fullname: "Biological Physics", position: 36, subscriptions_count: 12, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.chem-ph", source: "arxiv", fullname: "Chemical Physics", position: 37, subscriptions_count: 6, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.class-ph", source: "arxiv", fullname: "Classical Physics", position: 38, subscriptions_count: 3, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.comp-ph", source: "arxiv", fullname: "Computational Physics", position: 39, subscriptions_count: 13, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "physics.data-an", source: "arxiv", fullname: "Data Analysis, Statistics and Probability", position: 40, subscriptions_count: 11, last_paper_date: "2014-05-12 01:00:00", parent_uid: "physics"},
+  {uid: "math.CT", source: "arxiv", fullname: "Category Theory", position: 58, subscriptions_count: 12, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.AC", source: "arxiv", fullname: "Commutative Algebra", position: 61, subscriptions_count: 12, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.FA", source: "arxiv", fullname: "Functional Analysis", position: 65, subscriptions_count: 13, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.GM", source: "arxiv", fullname: "General Mathematics", position: 66, subscriptions_count: 14, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"},
+  {uid: "math.GN", source: "arxiv", fullname: "General Topology", position: 67, subscriptions_count: 7, last_paper_date: "2014-05-12 01:00:00", parent_uid: "math"}
 ])
