@@ -1,66 +1,115 @@
 FactoryGirl.define do
   factory :feed do
-    sequence(:uid)        { |n| "feed.#{n}" }
-    sequence(:fullname)       { |n| "Feed #{n}" }
-    sequence(:source)     { |n| "arxiv" }
-    last_paper_date       Date.today
+    sequence(:uid)      { |n| "feed.#{n}" }
+    sequence(:fullname) { |n| "Feed #{n}" }
+    sequence(:position) { |n| n }
+    source              'arxiv'
+    last_paper_date     { Date.today }
   end
 
   factory :user do
-    sequence(:fullname) { |n| "Person #{n}" }
-    sequence(:username) { |n| "person_#{n}" }
-    sequence(:email)    { |n| "person_#{n}@example.com"}
-    sequence(:active)   { |n| true }
-    account_status User::STATUS_USER
-    password "foobar"
-    password_confirmation "foobar"
+    sequence(:fullname)   { |n| "Person #{n}" }
+    sequence(:username)   { |n| "person_#{n}" }
+    sequence(:email)      { |n| "person_#{n}@example.com"}
+    active                true
+    account_status        User::STATUS_USER
+    password              'foobar'
+    password_confirmation 'foobar'
+
+    factory :admin do
+      account_status User::STATUS_ADMIN
+    end
+
+    trait :moderator do
+      account_status User::STATUS_MODERATOR
+    end
   end
 
   factory :category do
+    paper
     feed
-    sequence(:position) { |n| n}
+    sequence(:position) { |n| n }
+  end
+
+  factory :version do
+    paper
+    date { Date.today }
+    size "16kB"
+    sequence(:position) { |n| n }
   end
 
   factory :paper do |p|
     sequence(:title)       { |n| "On Hilbert's #{n}th Problem" }
     sequence(:abstract)    { |n| "We solve Hilbert's #{n}th problem." }
-    sequence(:submitter)   { |n| "Hilbert N. Grande" }
-    sequence(:uid)  { |n| "#{1000+n}/#{1000+n}.#{1000+n}" }
-    sequence(:submit_date) { |n| Date.today }
-    sequence(:update_date) { |n| Date.today }
+    submitter              'Hilbert N. Grande'
+    sequence(:uid)         { |n| "#{1000+n}/#{1000+n}.#{1000+n}" }
+    submit_date            { Date.today }
+    update_date            { Date.today }
+    pubdate                { Paper.estimate_pubdate(Date.today) }
     sequence(:abs_url)     { |n| "http://arxiv.org/abs/#{1000+n}/#{1000+n}.#{1000+n}" }
     sequence(:pdf_url)     { |n| "http://arxiv.org/pdf/#{1000+n}/#{1000+n}.#{1000+n}" }
-    sequence(:author_str)  { |n| "Hilbert N. Grande, Lucrezia Mongfish" }
+    author_str             'Hilbert N. Grande, Lucrezia Mongfish'
+
+    trait :with_comments do
+      after(:create) do |paper, evaluator|
+        create_list(:comment, 3, paper: paper)
+        create(:deleted_comment, paper: paper, content: "this is a deleted comment")
+      end
+    end
+
+    trait :with_categories do
+      after(:create) do |paper, evaluator|
+        create(:category, paper: paper)
+      end
+    end
+
+    trait :with_versions do
+      after(:create) do |paper, evaluator|
+        create_list(:version, 3, paper: paper)
+      end
+    end
+
+    trait :with_claimants do
+      after(:create) do |paper, evaluator|
+        create_list(:authorship, 3, paper: paper)
+      end
+    end
 
     factory :paper_with_authors do
       after(:create) do |paper, evaluator|
-        FactoryGirl.create_list(:author, 3, paper: paper)
+        create_list(:author, 3, paper: paper)
       end
     end
 
-    factory :paper_with_comments do
-      after(:create) do |paper, evaluator|
-        FactoryGirl.create_list(:comment, 3, paper: paper)
-        FactoryGirl.create(:comment, paper: paper, deleted: true, 
-                           content: "this is a deleted comment")
-      end
-    end
-
-    after(:create) do |paper, evaluator|
-      create(:category, paper: paper, feed: Feed.first)
-    end
+    factory :paper_with_versions, traits: [:with_versions]
+    factory :paper_with_comments, traits: [:with_comments]
+    factory :paper_with_categories, traits: [:with_categories]
+    factory :paper_with_comments_and_categories, traits: [:with_comments, :with_categories]
   end
 
   factory :author do
     paper
-    sequence(:position)   { |n| n }
-    sequence(:fullname)   { |n| "Lucrezia Mongfish" }
-    sequence(:searchterm) { |n| "Mongfish_L" }
+    sequence(:position) { |n| n }
+    fullname   "Lucrezia Mongfish"
+    searchterm "Mongfish_L"
+  end
+
+  factory :authorship do
+    paper
+    user
   end
 
   factory :comment do
     user
     paper
-    sequence(:content)  { |n| "This is test comment #{n}!" }
+    sequence(:content) { |n| "This is test comment #{n}!" }
+
+    factory :hidden_comment do
+      hidden true
+    end
+
+    factory :deleted_comment do
+      deleted true
+    end
   end
 end
