@@ -1,19 +1,12 @@
 require 'spec_helper'
 
 describe "Static pages" do
-
-  subject { page }
-
-  shared_examples_for "all static pages" do
-    it { should have_heading heading }
-    it { should have_title page_title }
-  end
-
   describe "About page" do
     before { visit about_path }
-    let(:heading)    { 'About' }
-    let(:page_title) { 'About' }
 
-    it_should_behave_like "all static pages"
+    it "has the right content" do
+      page.should have_heading "About"
+      page.should have_title "About"
+    end
   end
 end
