@@ -1,8 +1,30 @@
+# Code coverage for Travis
+require 'coveralls'
+Coveralls.wear!
+
+# OmniAuth test mode
+require 'omniauth'
+OmniAuth.config.test_mode = true
+
 require 'capybara/rspec'
 
 ENV["RAILS_ENV"] ||= 'test'
 require File.expand_path("../../config/environment", __FILE__)
 require 'rspec/rails'
+
+Delayed::Worker.delay_jobs = false
+
+# http://stackoverflow.com/questions/8774227/why-not-use-shared-activerecord-connections-for-rspec-selenium
+class ActiveRecord::Base
+  mattr_accessor :shared_connection
+  @@shared_connection = nil
+
+  def self.connection
+    @@shared_connection || retrieve_connection
+  end
+end
+
+ActiveRecord::Base.shared_connection = ActiveRecord::Base.connection
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
@@ -11,6 +33,8 @@ Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
 # Checks for pending migrations before tests are run.
 # If you are not using ActiveRecord, you can remove this line.
 ActiveRecord::Migration.check_pending! if defined?(ActiveRecord::Migration)
+
+Capybara.javascript_driver = :webkit
 
 RSpec.configure do |config|
   config.include Capybara::DSL
@@ -35,14 +59,15 @@ RSpec.configure do |config|
   # rspec-rails.
   config.infer_base_class_for_anonymous_controllers = false
 
-  config.before(:each) { reset_email }
+  config.before(:each) do
+    reset_email
+  end
 
   # Ensure database is cleaned properly
   config.before(:suite) do
+    Search.drop
     Search.migrate
-    DatabaseCleaner.strategy = :transaction
     DatabaseCleaner.clean_with(:truncation)
-    DatabaseCleaner.start
-    load "#{Rails.root}/db/seeds.rb" 
+    load "#{Rails.root}/db/seeds.rb"
   end
 end
